feat: settings.local.json生成ロジックを共通化 + 手動生成ボタン追加
worktree-manager内のインライン実装をsrc/lib/claude-settings.tsに抽出し、
タスク詳細画面に「Add Local Settings」ボタンを追加して手動で再生成可能にした。
